Query formats and servis implementations (in progress)

+++ src/tr/com/bookcell/favourite/DefaultFavouriteService.java
@@ -2,6 +2,10 @@ package tr.com.bookcell.favourite;
 
 import tr.com.bookcell.book.*;
 
+import java.util.List;
+
+import static tr.com.bookcell.util.InputFormatter.capitalizeFirst;
+
 public class DefaultFavouriteService implements FavouriteService {
     private final FavouriteRepository favouriteRepository;
 
@@ -11,21 +15,43 @@ public class DefaultFavouriteService implements FavouriteService {
 
     @Override
     public void add(Integer customerId, String bookName, String authorName, String authorSurname) {
+        String formattedBookName = capitalizeFirst(bookName);
+        String formattedAuthorName = capitalizeFirst(authorName);
+        String formattedAuthorSurname = capitalizeFirst(authorSurname);
         BookRepository defaultBookRepository = new DefaultBookRepository();
         BookService defaultBookService = new DefaultBookService(defaultBookRepository);
         Book book = new Book();
-        book = defaultBookService.getByNameAndAuthor(bookName, authorName, authorSurname);
-
+        book = defaultBookService.getByNameAndAuthor(formattedBookName, formattedAuthorName, formattedAuthorSurname);
         Favourite favourite = new Favourite(customerId, book.getId());
-        favouriteRepository.add(new Favourite(customerId, book.getId()));
+        for (Favourite tempFavourite : getByCustomerId(customerId)) {
+            if (tempFavourite.getBookId().equals(favourite.getBookId())) {
+                System.out.println("THERE IS ALREADY " + formattedBookName + " IN FAVOURITES LIST");
+                break;
+            }
+            favouriteRepository.add(new Favourite(customerId, book.getId()));
+        }
     }
 
     @Override
     public void remove(Integer customerId, String bookName, String authorName, String authorSurname) {
+        String formattedBookName = capitalizeFirst(bookName);
+        String formattedAuthorName = capitalizeFirst(authorName);
+        String formattedAuthorSurname = capitalizeFirst(authorSurname);
         BookRepository defaultBookRepository = new DefaultBookRepository();
         BookService defaultBookService = new DefaultBookService(defaultBookRepository);
         Book book = new Book();
-        book = defaultBookService.getByNameAndAuthor(bookName, authorName, authorSurname);
-        favouriteRepository.remove(customerId, book.getId());
+        book = defaultBookService.getByNameAndAuthor(formattedBookName, formattedAuthorName, formattedAuthorSurname);
+        for (Favourite tempFavourite : getByCustomerId(customerId)) {
+            if (tempFavourite.getBookId().equals(book.getId())) {
+                favouriteRepository.remove(customerId, book.getId());
+                break;
+            }
+            System.out.println("THERE IS NO " + formattedBookName + " IN FAVOURITES LIST!");
+        }
+    }
+
+    @Override
+    public List<Favourite> getByCustomerId(Integer customerId) {
+        return favouriteRepository.getByCustomerId(customerId);
     }
 }
